Make deeply nested objects update localStorage
This ended up being a much larger PITA than I would've thought. A
circular reference seemed to magically appear on sets of nested objects
that ended up somehow being caused by the fact that I passed along the
receivers from Proxy traps. I just find + replaced any mention of
receiver and got rid of it, and that seemed to do the trick.

In case you're wondering about how I discovered that this was the
problem throughout this nonsense: it turns out that the receiver that
used to be passed along in line 318 of index.ts in this commit was
actually a Proxy, even though the target was not. This probably happened
due to my manual calling of parent proxies' set traps. In any case, no
more problems for now.