Change how modify works for storeSeparate
The modify function is now passed a partial version of the full object
passed to storeSeparate, containing only the relevant key. The modify
function should return a similar partial object, where only the key
passed to it is acknowledged and modified.

I chose to do it this way to keep some form of type inference instead of
just letting any value be returned from modify.