feat: Add factories for special stores to library
Adds a subfolder for factory functions that return special stores. To
start, the only function is one that stores a value in React's state
instead of in localStorage.

It would be a good idea to update the docs (maybe even rename the
library at some point?) to make it clear that this is a feature.